Added graphs using Google Visualization API
Graphs are hard and Google is silly. Don't do anything with the graphs
until AFTER it is loaded, and must have the load call in a timeout so it
loads after the DOM (otherwise google will do a document.write, nuking
the whole page). Added mana color distribution graph to builder.